# Conflicts:
#	flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/out/persistence/assessment/AssessmentPersistenceJpaAdapter.java
#	flickit-assessment-data/src/main/java/org/flickit/assessment/data/jpa/users/spaceuseraccess/SpaceUserAccessJpaRepository.java

+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/out/persistence/assessment/AssessmentPersistenceJpaAdapter.java
@@ -13,6 +13,7 @@ import org.flickit.assessment.data.jpa.core.assessment.AssessmentJpaEntity;
 import org.flickit.assessment.data.jpa.core.assessment.AssessmentJpaRepository;
 import org.flickit.assessment.data.jpa.core.assessment.AssessmentListItemView;
 import org.flickit.assessment.data.jpa.core.assessmentresult.AssessmentResultJpaRepository;
+import org.flickit.assessment.data.jpa.kit.question.QuestionJpaRepository;
 import org.flickit.assessment.data.jpa.kit.assessmentkit.AssessmentKitJpaEntity;
 import org.flickit.assessment.data.jpa.kit.assessmentkit.AssessmentKitJpaRepository;
 import org.flickit.assessment.data.jpa.kit.maturitylevel.MaturityLevelJpaEntity;
@@ -29,7 +30,7 @@ import java.util.*;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static org.flickit.assessment.core.common.ErrorMessageKey.GET_ASSESSMENT_PROGRESS_ASSESSMENT_RESULT_NOT_FOUND;
+import static org.flickit.assessment.core.common.ErrorMessageKey.*;
 import static org.flickit.assessment.data.jpa.core.assessment.AssessmentJpaEntity.Fields.ASSESSMENT_KIT_ID;
 import static org.flickit.assessment.data.jpa.core.assessment.AssessmentJpaEntity.Fields.SPACE_ID;
 import static org.springframework.data.jpa.domain.Specification.where;
@@ -43,13 +44,13 @@ public class AssessmentPersistenceJpaAdapter implements
     GetAssessmentProgressPort,
     GetAssessmentPort,
     DeleteAssessmentPort,
-    CheckAssessmentExistencePort,
     CountAssessmentsPort,
     CheckUserAssessmentAccessPort {
 
     private final AssessmentJpaRepository repository;
     private final AssessmentResultJpaRepository resultRepository;
     private final AnswerJpaRepository answerRepository;
+    private final QuestionJpaRepository questionRepository;
     private final AssessmentKitJpaRepository kitRepository;
     private final SpaceJpaRepository spaceRepository;
     private final MaturityLevelJpaRepository maturityLevelRepository;
@@ -139,6 +140,9 @@ public class AssessmentPersistenceJpaAdapter implements
 
     @Override
     public UpdateAssessmentPort.Result update(UpdateAssessmentPort.AllParam param) {
+        if (!repository.existsByIdAndDeletedFalse(param.id()))
+            throw new ResourceNotFoundException(UPDATE_ASSESSMENT_ID_NOT_FOUND);
+
         repository.update(
             param.id(),
             param.title(),
@@ -150,28 +154,27 @@ public class AssessmentPersistenceJpaAdapter implements
     }
 
     @Override
-    public GetAssessmentProgressPort.Result getAssessmentProgressById(UUID assessmentId) {
+    public GetAssessmentProgressPort.Result getProgress(UUID assessmentId) {
         var assessmentResult = resultRepository.findFirstByAssessment_IdOrderByLastModificationTimeDesc(assessmentId)
-            .orElseThrow(() -> new ResourceNotFoundException(GET_ASSESSMENT_PROGRESS_ASSESSMENT_RESULT_NOT_FOUND));
+            .orElseThrow(() -> new ResourceNotFoundException(GET_ASSESSMENT_PROGRESS_ASSESSMENT_NOT_FOUND));
 
         int answersCount = answerRepository.getCountByAssessmentResultId(assessmentResult.getId());
-        return new GetAssessmentProgressPort.Result(assessmentId, answersCount);
+        int questionsCount = questionRepository.countByKitVersionId(assessmentResult.getKitVersionId());
+        return new GetAssessmentProgressPort.Result(assessmentId, answersCount, questionsCount);
     }
 
     @Override
     public Optional<Assessment> getAssessmentById(UUID assessmentId) {
-        Optional<AssessmentJpaEntity> entity = repository.findById(assessmentId);
+        Optional<AssessmentJpaEntity> entity = repository.findByIdAndDeletedFalse(assessmentId);
         return entity.map(AssessmentMapper::mapToDomainModel);
     }
 
     @Override
     public void deleteById(UUID id, Long deletionTime) {
-        repository.delete(id, deletionTime);
-    }
+        if (!repository.existsByIdAndDeletedFalse(id))
+            throw new ResourceNotFoundException(DELETE_ASSESSMENT_ID_NOT_FOUND);
 
-    @Override
-    public boolean existsById(UUID id) {
-        return repository.existsByIdAndDeletedFalse(id);
+        repository.delete(id, deletionTime);
     }
 
     @Override
